requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
-e git+https://github.com/ioet/bpm-projects-api.git@e87f5901e3f065d817427af73471036b483d9fad#egg=bpm_projects_api
cffi==1.11.5
click==6.7
cryptography==2.3.1
feedparser==5.2.1
Flask==1.0.2
flask-restplus==0.11.0
Flask-Script==2.0.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
meds==1215
more-itertools==4.3.0
pluggy==0.7.1
project==20
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyJWT==1.6.4
pytest==3.8.0
pytz==2018.5
six==1.11.0
sleekxmpp==1.3.3
Werkzeug==0.14.1
